requirements.txt
# pip install -r requirements.txt
python-dotenv
pymongo
flask-bcrypt
marshmallow
pyjwt